hw4/requirements.txt
Cython==0.25.2
Keras==2.0.2
Pillow==4.1.0
PyYAML==3.12
Pygments==2.2.0
Theano==0.9.0
appdirs==1.4.3
args==0.1.0
astroid==1.5.2
boto==2.46.1
bz2file==0.98
chardet==2.3.0
clint==0.5.1
colorama==0.3.3
coverage==4.3.4
cycler==0.10.0
decorator==4.0.11
distlib==0.2.1
ez-setup==0.9
gensim==2.0.0
graphviz==0.6
h5py==2.7.0
html5lib==0.999
imageio==2.1.2
ipython==6.0.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
keras-vis==0.2
lazy-object-proxy==1.3.1
matplotlib==2.0.0
mccabe==0.6.1
numpy==1.12.1
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pkginfo==1.4.1
prompt-toolkit==1.0.14
protobuf==3.2.0
ptyprocess==0.5.1
py==1.4.33
pydot==1.2.3
pydot-ng==1.0.0
pylint==1.7.1
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.4.0
python-apt==0.9.3.11
python-dateutil==2.6.0
pytz==2017.2
requests==2.8.1
requests-toolbelt==0.7.1
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
smart-open==1.5.2
tensorflow-gpu==1.0.1
traitlets==4.3.2
twine==1.8.1
typing==3.6.1
unattended-upgrades==0.1
urllib3==1.12
versioneer==0.18
virtualenv==1.11.6
wcwidth==0.1.7
wheel==0.29.0
word2vec==0.9.1
wrapt==1.10.10
yapf==0.16.1
